Updates Model initialization paradigm - now specify labels & basis.
Previously, you could initialize a Model (or GateSet) without
specifying any state-space-labels or a basis.  This was fine, as
labels weren't really needed (except for embedded ops) and neither
was the basis (except for Lindblad ops and postprocessing I think).
The required bits - the dimension of the model and its 'evotype' -
were just determined by the first object that was inserted.

State space labels and a 'matrix' basis is more needed by a model
now, and since it's easy to specify these things in the 1- and 2-
qubit cases even though they're not always needed, this commit
changes Model initialization to *require* state-space-labels, a
basis, and an evotype.  (From the labels or basis we can compute
a dimension.)

This removes some convenience, but improves code clarity.  Furthermore,
it simplifies OrderedMemberDict, which previously had to *set* its
parent's .dim (now just a read-only property) or ._evotype when these
where set to None.

This commit updates most of the model-creation occurrances in the
package code to supply the new needed arguments, but not in the
unit tests (though I don't think there's many instances there since
models are usually created using a construction function).

A couple tutorial notebooks are also renamed.