Updates Basis objects to compute matrices only when needed (lazy).
This way, we can create Basis objects for large numbers of qubits
without actually allocating & creating the matrices, as we shouldn't
ever need them.  As a further safeguard, this commit adds a check
which raises a ValueError if the creation of basis matrices w/dim
greater than pygsti.baseobjs.basisconstructors.MAX_BASIS_MATRIX_DIM
is attempted.