Adds line labels to Circuit string rep, e.g. "Gx:1Gy:2@(0,1,2)"
Now the single-line string representation of a Circuit object also
contains its line labels, so it can be properly serialized & de-
serialized using this representation (e.g. in text format files).
For simplicity and backward compatibility, if the Circuit's line
labels == ('*',) then no "@(...)" is appended to the string rep,
and likewise if a circuit is loaded from a string that doesn't
contain a "@(...)" line label specification, then the line labels
are automatically determined (so == ('*',) if there are no gate
labels with non-None sslbls and otherwise just those sslbls that
are present).  Lightly tested.

This commit also changes the default to `check=True` in Circuit.__init__
so that by default string representations *are* checked for validity.
A similar check is always performed when setting the .str property
of a Circuit now.  We explicitly set check=False in routines where
we've just loaded a circuit from a file/string and in internal
"trusted" routines such as Circuit.__add__.