Updates LGST to work with new ExplicitOpModel.__init__ signature.
This was mostly updated in a prior commit, but the LGST model was
then *always* initialized with the target model's state space
labels and basis.  In the case when we've set svdTruncateTo we
might have a different dimension state space and basis, so we need
to create a "dummy" state space labels and basis.  This commit adds
this logic, creating a dummy space & basis for a k-dimensional
superop space (superops are k x k matrices) interpreted as the
direct sum of k 1-D blocks on the diagonal of a k x k "density matrix".

This commit performs a similar update to ExplicitOpModel's
increase_dimension and decrease_dimension, which always need dummy
spaces (for the same reason: that you can't simply use the old/
target space labels & basis when the model dimension changes).