Updates ImplicitModel to support "categories" of operators.
Now the operator-holding members of ImplicitModel (i.e.
prep_blks, operation_blks, povm_blks and instrument_blks) are
*dictionaries* of OrdereMemberDict objects instead of just a
single OrderedMemberDict.  The outer dict holds "categories" of
operators such as "layers" (the default, if these is such a thing),
"gate", etc.  LocalNoiseModel creates "layers" and "gates" operation
categories and CloudNoiseModel creates "layers", "gates", and
"cloudnoise".  Both create just a "layers" category preps, povms,
and instruments for now.

This makes it easier to see which operators are used/intended for
different purposes within an implicit model, and avoids having to
place a "category_" prefix in the Labels themselves, which was done
previously and is rather kludgy.

Relevant unit tests updated too, so all-but-drift tests should
continue to pass.

Note: this commit also fixes a bug in slicetools.shift: the
"null slice" slice(0,0,None) should *not* be shifted (and now
isn't).