Updates various tool functions so reports with multi-block Bases work.
"work" is generous here - this commit makes it possible to create
a HTML report using a model whose .basis has multiple tensor-product-
blocks (e.g. for leakage modeling).  It still doesn't work with
everything, in particular we punt on a couple reportables.py
functions where we just return -1 if the basis has >1 blocks.