Moderate renaming: "compile" circuit (for fwdsim) -> "simplify" circuit
This commit performs an automated renaming of "compile" to "simplify"
when we're referring to the process whereby we prepare a circuit
(containing a POVM and possibly instruments) for use by a forward
simulator (which just deals with preps, ops, and effects).  This
process was called circuit "compiling", but because there's an
obvious naming collision with the process of changing the primitive
gates in a circuit (e.g. a Clifford "compiler"), we've chosen to
now call the aforementioned process "circuit simplification" (not
currently used in pyGSTi).  This commit updates the code and unit
tests, but the unit tests haven't been rerun yet, so expect minor
fixes to follow.  The commands used to perform the bulk of the
renaming were (the only repairs from this were from the .cos -> .sos
replacement, which hit a few _np.cos calls accidentally):

Main code: **Note** this line was NOT run on the files compilationlibrary.py,
 compilers.py, or processorspec.py as these were removed from the temporary
 source (original) replacement directory.
> python3 ../replacer/doReplace.py ../replacer/update4_replacements/repl.txt ../pygsti/ *.py */*.py

Unit tests (again from a temporary sibling directory):
> python3 ../../packages/replacer/doReplace.py ../../packages/replacer/update4_replacements/repl_for_tests.txt ../test_packages/ *.py */*.py