Fixes a basis-bug in ProjectionsBoxPlot.
Fixes bug that when a Basis object is given as the projection basis
AND we reshape the input array, e.g. from a (16,1) -> (4,4) shape,
the *name* of the projection basis must be used to create the smaller
(e.g. 4-dimensional) bases used for label generation NOT the entire
Basis object (which will remain 16-dimensional even when a 4-dim
basis is requested).