Sets the choiMxBasis == mxBasis in entangelement fidelity calculations.
Previously only the mxBasis was set when performing a Choi-Jamiol.
isomorphism.  Since we (recently) changed the default bases to "pp"
the infideltiy function would fail when using a gm/qutrit basis
on a non 4^n-dimensional HS space.  This commit sets the choiMxBasis
to whatever the mxBasis is, which should fix this issue. (We could
probably also have just used the 'gm' basis all the time, since
the choi matrices just get used to compute & return a fidelity).