Fixes bug in fastreps.cpp within EmbeddedOp's apply method.
There was an error in applying the trivial action of the non-active
tensor-product-blocks for an embedded operation.  This bug was only
in the C-version of the action, and only would have caused a problem
if the state space contained multiple tensor product blocks (which
I don't it ever did). Bug was found by re-inspecting the code.