Updates Basis.__eq__ to avoid generating matrices.
Sometimes, like in the initialization of a composed gate, we
check to make sure a number of bases are the same using the ==
operator on Basis objects.  This used to cause the basis matrices
to be build (so they could be compared) but this needlessly uses
a ton of memory when the bases are large.  Now __eq__ just compares
the name, dim, and sparsity (the things used to lazily generate
the basis matrices) of Basis objects when the matrices haven't be
built yet.