Fixes pickling error cause by increased caching (prior commit).
Caching _computeSubMxs causes numpy object arrays to be stored
in the cache.  These are fine for pickles, but they can't be
serialized as json/msgpack because np.fromstring won't work for
object arrays.  This update causes the json coded to throw a
TypeError when attempting to serialize an object ndarray, which
the added SmartCache.__pygst_getstate__ catches and skips when
it serializes itself.

This commit also adds get/set state functions to ReportFigure,
since this class contains a plotly figure which can have issues
with *un*-pickling (so we call .to_dict() and pickle that).  I
don't think this was actually causing a problem anywhere, but it
seems needed to properly pickle ReportFigure objects.