Fixes bug with diamond-norm error bars.
After the update to CVXPY 1.0 the 'evaluate_nearby' method of
diamond-norm classes in reportables.py broke (a while ago) because
cvxpy Variables don't multiply as matrices using the '*' operator
anymore.  This update may break compatibility with early-version
CVXPY (if the old Variables are incompatible with np.dot) - this
hasn't been tested.