Revamped Basis object for  state-vector and classical bit compatibility.
This significant commit restructures the Basis object as being a
base class for the new BuiltinBasis, ExplicitBasis, TensorProdBasis,
and DirectSumBasis classes.  The way builtin bases are created in
basisconstructors.py has also changed significantly.  The biggest
outward change is that we're trying to do away with the "Dim" object,
and bases now just hold as their ".dim" what was their operator
dimension - the dimension of the vector space they serve as a full
or partial basis for.  The new, more structured object hierarchy
has in view the ability to have vector-type bases for state vectors
and classical bits.  A Basis object has .elshape and .elsize members
which give the shape and (flattened) size of the object's .elements
array (computed lazily in some cases).  Note that .elsize does not
necessarily equal .dim, as in a DirectSumBasis the .elsize refers
to the dimension of the "embedded space" whereas .dim only counts
the kite-structure subspace.

Many related modules are updated to adjust to the change in Basis
object. There's still more modifications that are needed, and some
modifications are just hacks for now (see optools.py) -- search for
"HACK" and "FIX" to find marked locations.

Also, much of the old Basid code has been commented out for reference
and marked with "REMOVE", to be removed in the future.