Fixes lingering "Gi" error in create_cloudnoise_sequences.
We removed the ability to recognize the special "Gi" layer from
LocalNoiseModel and CloudNoiseModel a little while ago but forgot
to update create_cloudnoise_sequences(...).  Therein, we assume
the idle string is Circuit(("Gi",)) - which now causes problems
down the line since the relevant Models don't recognize this
circuit/layer.  This commit fixes this issue by making the idle
string an argument, `idleOpStr`, to create_cloudnoise_sequences
that defaults to the way we indicate idle layers now: ((),)