Further updates unit test files given updated Basis and StateSpaceLabels.
Now nearly all of the unit tests pass, and the ones the fail are
either expected to fail or are understood.  The biggest uncertainty
now regards the hard-coded "answer" for the CnotA gate in the
helper_constructGates function of testGateSetConstruction.py --
it seems that the hardcoded answer may have been incorrect.