Adds (untested) ability to unpickle v0.9.7 objects using legacyio.py
We need to test this, but to do so we'll need to create some test
files using another branch, and so need to commit this now.  This
commit sets up a paradigm whereby legacyio.py will be able to support
loading files from various old versions, via a new argument to the
enable_old_object_unpickling context manager function.