Adds better unpickling support for 0.9.7.1 types.
In particular allows Circuits to be updated to exponentials-as-blocks
representations (by looking *only* at the stored string reps - something
we may want to change in the future), and adds needed upgrade conversion of
StateSpaceLabels objects.

See Circuit_setstate() and Hack_CompressedCircuit_expand() in
legacyio.py to see how circuits are read in based only on their
string reps.