Adds commented-out assert statements for typical too-high minprobclip error.
Within the logL objective functions in do_mlgst, these assert
catch an error that manifests itself as a "objective function as
infinite value at initial point" error.  This error typically occurs
because we take the square root of a negative number, which in turn
occurs because the "minProbClip" is set to a value larger than the
minimum observed non-zero frequency or predicted probability.  Since
this assert statement checks an entire array for negative elements
and is in the innermost loop of GST optimizations, we've left it
commented out for now.