Fixes Circuit.str being out-of-sync with labels.
Now Circuit.str regenerates its string representation whenever
the Circuit is editable, fixing issues whereby the string representation
gets set/cached and then circuit is changed w/out refreshing the
_str member.  When a Circuit is static, it will lazily cache its
generated string representation since it can't be changed anymore.
If the user assigns a Circuit's .str attribute, this is assumed
to be correct and is not re-generated, even when the Circuit is
editable.

This commit also fixes a couple Circuit bugs:
- use of undefined variable in Circuit.__mul__
- nicer printing of editable circuits (previously an editable
  Circuit's .str might get auto-generated as [Label(Gx),Label(Gy)]
  b/c lists weren't handled properly).