Updates map-based evaluation tree for better performance.
Adds plumbing to propagate a "maxCacheSize" down from Model.set_simtype(...)
to the creation of a MapEvalTree, which now treats a max-size of
zero specially, as this makes it much easier to create & split
trees.  Included in this is a new argument, `all_final` to
EvalTree._finish_split(...) and a slight change in the way evaluation
trees are created: previously Model would call a forward sim's
construct_evaltree function and then call evTree.initialize(...);
now all the initialization is done by the forward sim, so that
forward-sim arguments (like maxCacheSize) can be used in tree
construction.  As a result, construct_evaltree takes more arguments.