Bug fixes to Circuit.
Makes default_expand_subcircuits a member of Circuit, and fixes
bug caused by using this as the default value of `expand` in
Circuit.__init__.  Now expand defaults to "default", which uses
the value of Circuit.default_expand_subcircuits.

Also, this fixes a bug in Circuit.serialize() whereby an empty
layer was serialized as the empty string, which isn't correct.
Now the empty layer is treated as a special case that is serialized
as itself (an empty layer). Note: this fixes the behavior of
create_cloudnoise_sequences(...).