Adds limited support for CircuitLabels in local- & cloud-noise models.
The layer lizard classes in cloudnoisemodel.py and localnoisemodel.py
now use a (new) common base-class method, "get_circuitlabel_op", to
generate an operation for a CircuitLabel object.  In the case of
CloudNoiseModel objects, this will work for "top-level" subcircuits
which act on the entire state space, but this won't work (and it's
unclear exactly what the desired output is - so now a NotImplementedError
is raised; see the comment in cloudnoisemodel.py near the 'raise'
statement) when there are subcircuits *within* layers or other subcircuits.