Fixes 'resize_std_mx' basistools function.
In Jamiolkoski tools (and perhaps other places), this resize function
is used to both change bases and resize if necessary, e.g. from
a standard basis (which may be a tensor product of std components now)
to a "simple" standard basis (which can't be a tensor product).
The previous implementation did not perform a basis change when
the basis dimensions where the same, which was find when we didn't
have tensorprod bases.  This commit adds a fix so that the function
performs a basis change even when the basis dimensions are the same
(and no resizing is needed).  Note: maybe we should just use
flexible_change_basis(...) for this??