Another tensor-product basis related bug fix.
Replaces "std" with matrix_basis.simple_equivalent("std") in
construction of error generators (in _init_generators(...)), as
now changing bases from a tensor product Basis to "std" will
convert to the std basis that is a tensor product of std components,
which is not what we want.  Using simple_equivalent works, as
would just creating a builting "std" basis of the appropriate
dimension.