Improved tensor-product basis bug fix.
Improves the last commit in terms of code clarity and perhaps
future functionality: replaces matrix_basis.simple_equivalent("std")
with the creation of a BuiltinBasis of type "std" and of the
appropriate dimension.  This is preferable because simple_equivalent
"flattens" direct sums, which we don't really need here - we just
need a non-tensor-product "std" basis.