Fixes basis labels in error-gen projections plots for 2Q GST.
TensorProdBasis objects don't have a nice .name (like 'pp' or 'gm'),
and so the logic for finding the x- and y-axis bases for a 2Q
error-generator-projection plot was failing to recognize that a
TensorProdBasis with 2 components (1Q 'pp' basis x 1Q 'pp' basis)
had the x- and y-axis label bases contained in it.  This commit
adds this as a special case.