Changes Circuit.get_layer to return the *components* of the layer label.
If c is a Circuit, c.get_layer(j) originally did the same thing as
c[j], which is to return a layer Label.  This may be confusing,
as one might expect an iterable of gate labels to be returned, and
so now get_layer returns the *components* of the layer and the new
get_layer_label function retains the original behavior.  This commit
updates all get_layer usage in pyGSTi to get_layer_label, so nothing
should be broken by this.