Adds basic wildcard allocation within do_mlgst.
This is just in the testing/debugging stages, but now there's a
flag (harcoded to True for now) that enables wildcard budget
allocation and can for our 1Q test find a 1-norm minimial
"wildcard-vector" that meets a total-LogL p-value criterion and
a no-red-boxes criterion simultaneously.  We're just using a
simple 1D wildcard vector whose element multiplies the length
of a circuit for testing.