Moves wildcard calculation to act as a "robust scaling" option.
Moves Wvec-finding algorithm from do_mlgst to the post processing
in longsequence.py.  Also updates the algorithm slightly, giving
Wvec a component for each primitive gate label in the model, and
starting the algorithm at a larger eta (this should be done better
in the future - the idea is that we want to start at a point
where the eta*|Wvec|_1 constraint is actually active).  We also
seed Wvec optimizations with the results of the prior step, instead
of always starting at a `Wvec_init`.