Updates unit tests to fix some TravisCI failures.
In particular, adds a HACK for unpicking in python 3.5 UUIDs created
in python 3.7 (which has a SafeUUID class).  This is in dataset.py
and meant to be a bandaid.  Other unit tests have been updated to
allow more tests to run with a failed cvxpy install, which seem to be
happening on TravisCI builds now.