Integrates wildcard budget into reports.
As a part of this integration, a new WildcardBudget object has
been created so we can easily pass information to plot and table
functions.  This, in turn, prompted updating the likelihoodfns.py
functions to accept a WildcardBudget object so that reporting
routines wouldn't need much updating.

The HTML report now has a new "Unmodeled errors" tab that shows
up when wildcard budgeting is enabled.  This has only been lightly
tested, but seems to work as desired for a simple 1Q case.