Adds "stages" to wildcard budget algorithm; cleans up some.
Now the search for a good "eta" value (strength of regularization
term) is done in two stages: 1) find a large enough eta that the
objective function is *very* positive (>100 now) - this gives us
a sensible starting point where we know the regularization term
is causing an effect.  2) reduce and refine eta to a value on the
"edge of the lake", as done previously.  We only restart from the
last solution in stage 2 when the objective function was positive,
since we don't trust the optimized Wvec when the objective function
is zero (could be anywhere on the lake!).

This commit also cleans up the algorithm code (in longsequence.py),
removing pieces that are unnecessary since we've migrated functionality
into the likelihoodfns.py functions.