Initial commit of feature-directterms branch.
This goal of this branch is to implement term-based calculations
using RankOneTerms with floating-point "direct" coefficients
rather than the polynomials used thus far.  This would eliminate
the need for a polynomial cache, and allow for on-the-fly path-integral
truncation (good things), but it's unclear what the cost of this
will ultimately be (maybe too slow to be useful?).

This initial commit sets up some basic framework for doing this
which is super slow:  the prs_directly(...) function in
termforwardsim.py computes probabilities using terms with normal
floating point coefficients (using a helper function only in
 slowreplib.py for now - so you can't use C extensions yet).  It
seems that there should be some lightweight caching we can do to
make this faster (as it's dog-slow even for 1Q now).  Note that
the way to enable the "direct terms" is to create a TermForwardSim
simulator with it's cache set to None (in model.py we don't supply
a default empty dict {} when the user-supplied cache is None anymore).