Adds Cython (C) support for direct (no polys) term-based calcs.
This adds the C-version of what was prototyped in the previous
commit in slowreplib.py, and tweaks things a bit more (adds
some caching) to make torep(...) calls faster and fewer (adds a
RepCacheEl object to fastreplib.pyx).  This may have broken the
slow version, but not badly.  In any case, the direct term-based
calcs run a bunch faster now, though maybe still too slow to be
useful (we'll see!).