Adds initial path-pruning logic to direct terms along w/debug stmts.
This commit adds logic to skip "paths" (sequences of error terms)
after a certain "path weight", relative to the total possible weight,
has been accounted for.  The debug statements can track that this
seems to be working (fewer paths are kept then less weight needs
to be accounted for), though it seems to be a pretty steep transition
from keeping none to nearly all of the paths.  The weight of each
term is held separately from its coefficient so that we can ensure
that the *same* paths are kept when computing finite-difference
derivatives (we only update the weights at the beginning of each
derivative calc).  Also, all 0-th and 1-st order terms are always
kept, so that we don't "lose traction" and compute derivatives that
are completely insensitive to a parameter (e.g. a param reaches a
fixed point of having a small value => none of the paths it contributes
to are included => there's no derivative to make us want to increase
the parameter).

On a simple 1Q 1-st-order term test we find that running with poly/direct
terms using either the python or Cython versions give the same
results (though vastly different run times!) - so it seems like the
pieces are correctly coded at this point (but w/lots of ugliness).

Run times are still long for the direct terms - I think the next
thing to try is using poly terms but computing them more on-the-fly
using the already-separate weights.

Note 1: the probabilities generated when using path
pruning are not strictly *real* because we don't prune in a way
that's guaranteed to preserve the imaginary-part cancellation that
occurs when we include all the paths.

Note 2: this is the last commit where the poly-term calcs have been
left unchanged since beginning work on direct terms & path pruning.
From here onward I expect we'll be trying to merge these two methods/
approaches into one to try to get the best of both worlds.

Still very much a WIP.