Continue development on truncated/pruned-path term calcs.
This commit adds a number of debugging statements and structures,
and fixes most of the bugs in the pruned path-integral calculation.
The main outstanding item is that high-magnitude terms should always
be ordered by magnitude: that some of the terms should always
be traversed at first order must be designated not by a number of
"first order always traversed" (foat) terms but by their indices,
since these terms need not appear in a contigous block of the index
space.

This commit also upgrades the Polynomial class so it can handle
higher order terms - instead of a single 64-bit int describing a
term of the polynomial, now a tuple of ints (to be a custom struct/
class in the C-version) compactly encodes the variable indices in
a Polynomial and a PolyRep (in slowreplib.py).

Note that everything so far uses slowreplib.py - no fast C implementation
has started yet because we're still trying to solidify the concepts
in the pure-python variant/version, i.e. in slowreplib.py.

This commit adds some term-rep and probability polynomial caching,
but it's still pretty messy.  Lots of debug statements everywhere.