Updates path traversal so terms in FOAT paths are designated by index.
As described in the prior commit, there is a need to keep all the
terms ordered my magnitude so that path traversal works efficiently
(so fastmode=True will work someday).  This is in tension with
designating some terms as "first order" such that the paths in
which they are the only non-zeroth-order term present are always
traversed.  Previously we just kept track of the number of 1st
order terms, but this can't work in conjuction with the sorting
we desire -- this commit introduces a new approach whereby we
keep track of the indices of the "1st order" terms for each op,
so they can appear anywhere within the ordered list of terms (and
still always be traversed).

Still lots of debug statements and checks.