Cleans up some debug statements & code - ready to begin Cython.
This commit comments out a bunch of debug statements and marks
them with TODO REMOVE for later removal.  Output is now fairly
concise so that one can run GST and see the outer iterations
clearly.  At this point, we're ready to start mirroring the slow
python versions of functions in Cython.