Adds Cython support for pruned-path term calcs.
Cython version runs and seems to roughly work, but hasn't been
compared in detail to Python version, and hasn't been optimized
much.  The "fastmode" is still disabled/unimplemnted in both
versions (but this isn't currently the bottleneck) since the
caching eliminates the need for much forward simulation in the
low-qubit tests we've done so far.