Adds fastmode=True support for Cython pruned terms & other optimizations.
The main update in this commit is the addition of fastmode=True
variants of the path-adding code in fastreplib.pyx.  This functionality
already existed in slowreplib.py but this commit enables it's use.
It has been tested (sucessfully) that the same 1Q result (to many
decimal places) is obtained whether fastmode=True/False is used
and whether the Python/Cython implementation is used, so it seems
that all four of these implementations are correct (or are wrong
in the exact same way).

This commit also makes some minor updates which try to optimize
the pruned-path term calcs:
1) Moves the "repcache" (a cache of TermReps) to be a member variable
 of TermEvalTree so that less recalculation of rep objects is needed.
2) adds a "recalc_threshold" argument to prs_as_pruned_polys that
 causes a recomputation of relevant paths (& assoc polynomials) only
 when computing a jacobian (in bulk_fill_dprobs).  This seems fine
 for use with the customLM algorithm but is a HACK that needs to
 be fixed in the future -- maybe we need to add some intelligence
 to the LM algorithm or at least a flag within the forward simulator
 to enable this behavior, as it would be surprising and potentially
 problematic when using the fwd simulator for other applications
 (e.g. if we never need to compute jacobians).