Adds some finishing touches on pruned-path calcs, preparing for develop merge.
Adds a default of zero (0) finite-difference iterations for
"termorder" and "termgap" simulation type (maybe we don't want this
for termorder??), as finite difference derivatives don't work for
termgap calcs.

Adds an "opt-mode" argument to TermForwardSim class which can put the
simulator into a mode whereby it only recomputes paths when bulk
probability derivatives are computed, as this seems reasonable for
use within an optimization.  This requires the user to specify a
":True" in the simtype, e.g.

mdl.set_simtype("termgap:3:0.0005:0.0005:True", calc_cache)

and could be improved in the future with an interface whereby the
customlm.py routines can inidicate what *type* of evaluation they're
asking for, beyond just an objective fn or jacobian.  Left as future
work.

Finally, this commit cleans up some debug statements so that only a
summary message regarding how many paths were chosen is printed.

At this point the pruned-path calculation code basically works and
is decently optimized.  It's been tested on 1 and 2-qubit cases,
and seems to work when the error rates are low enough and the sample
count is high enough.  Can be merged into a main branch and
tested on larger systems.