Adds fix for working with plotly v3.8 (released 6 hours ago).
Plotly 3.8 doesn't include the _plot_html function we use to
extract the raw html of a plotly plot; instead we get this from
the new plotly.io.to_html function (not a underscore-preceded one!).
The bad news is that we need to manually extract the UUID created
in to_html from the HTML string it returns - to_html(...) doesn't
return the plotly-div-id separately as _plot_html did.  Testing 1Q
example report generation indicates that figure display and resizing
still works as before after this fix, so I think we're good to go.