First commit on feature-modelbuilding branch.  Improves errgen coeffs.
Updates the L-term-dict, basis-dict pair of dictionaries for describing
the terms in an error generator in several important ways:
- the basis-dict has been replaced by an honest Basis object.  The
  EmbeddedBasis class was added to make embedding basis elements easy.
- scaling has been applied to the lindblad error generators (created
  by functions in lindbladtools.py) so that in a LindbladErrorGen object
  the Hamiltonian coefficients <=> angles and the Stochastic coefficients
  are to first order equal to the Stochastic error rates of the resulting
  (post-exponentiated) channel.  This makes these coefficients as intuitive
  and meaningful as possible, we think.
- the labels used to describe the basis elements in an L-term-dict are
  inherited from Basis objects, so that these labels are no longer
  meaningless integers but things like "X" or "Z:Q0" (the latter for
  and EmbeddedBasis).

A get_coeffs(...) method has been added to EmbeddedErrorGen and
ComposedErrorGen objects, so that basically any error generator object
can be queried for its coefficients in a meaningful way.  These functions
by default only return a L-term-dict, and have a `return_basis` argument
that defaults to False.

Cloudnoise model construction has been updated to create error-generator
bases with proper Pauli labels, so that calling get_coeffs() or
get_errgen_coeffs() (method of a LindbladOp) on the elements of a
CloudnoiseModel give meaningful output.