Adds set_coeffs(...) and set_errgen_coeffs(...) to relevant operators.
Now the user can set error-generator coefficients using a L-term-dict
similar to that returned by get_coeffs/get_errgen_coeffs.  Furthermore,
these methods can scale any stochastic/non-Hamiltonian coefficients
so that they're interpeted as error *rates* (of the single-Pauli
contribution to a depolarizaton channel) instead of exponentiated
coefficients, so that it's easier to set coefficients with a particular
channel error rate in mind (maybe this will be come a separate function
in the future?).