Begins branch for fixing sparse-data bug(s).  Kinda fixed now.
A bug was found whereby the chi2 and logl objective functions were
not computed correctly when there were "blanks" in a DataSet being
analyzed, i.e. "sparse data".  This was because chi2 and logl terms
corresponding to blank-outcomes were not computed (assumed to be
zero) and this is incorrect.  This is because the chi2 term is
still N(p-f)^2/p and the logl term is N*f*log(p) - N*p ( != 0
when f == 0).  Only the poissonPicture=False logl term ( N*f*log(p) )
is zero when f is.  We *have* to use the poissonPicture=True in
least-squares optimizations because this ensures the positive terms
needed for the algorithm, so just using poissonPicture=False isn't
really an option: we either need to store the zeros in the data set
(this is what hotfix-0.9.7.4 does) or somehow account for the blanks
in a data set.

This commit implements the latter - it reverts recordZeroCnt back to
`False` in DataSet.add_count_dict and adds a bunch of logic to the
core.py objective and jacobian functions so that blanks, or "omitted
data" is accounted for.  It does this in a way slightly different than
the way it woud be done if we included all the zero counts - we lump
all the omitted probability for a circuit, i.e. the quantity
1.0 - sum_of_present_probabilities, together and treat it as though
it was a *single* probability term, adding it's contribution to the
*first* present-outcome-probability for a given circuit.  This is
the same as the separate accounting when all the individual
probabilities corresponding to zero-frequences are not "clipped":
both the logl term and the chi2 term is just `N*p` when f==0 (linear!).
When there *is* clipping (p not in [minp,1-minp] for chi2 and p < radius
for logl), however, the penalty function is not linear and so there
will be a difference between the "sparse-data" and "non-sparse-data"
cases (this can happen more than you might think, b/c zero counts
often correspond to small probabilities).  This approach still seems
best, as there's nothing particularly "right" about either approach
(both "patch" the objective fn), and we don't want to separately
account for all the zero-frequency outcomes in the sparse-data
case: there could be a ton of them (e.g. a 20Q dataset with 100
clicks per sequence).  We can always make the
minProbClipForWeighting and radius parameters really small.

Things we still need to do surrounding this:
- we can remove the poissonPicture=False case for our objective
 functions - these are actually useless b/c they're incompatible
 with the least squares solver.
- the functions in chi2fns.py and likelihoodfns.py need to be
 updated as the ones in core.py have -- they're still incorrect
 for sparse data.
- it seems useful to have some "sparse" arguments to functions
 like generate_fake_data and load_dataset -- so we have access to
 the recordZeroCnts argument of add_count_dict.  Maybe even a way
 to convert a DataSet from sparse <-> dense representation?
- check the speed of the added core.py functions and optimize
 some of the omitted-probs code (maybe add some Cython fns,
 and at least short-circuit when there are no omitted probs).

Finally, this commit leaves a bunch of debugging code in core.py
- search for TODO REMOVE to get rid of this later.