Removes (comments-out) non-Poisson-picture mlgst case.
A long time ago we moved to a least-squares solver and, by default,
the "poisson picture" of the log-likelihood.  We left in the option
to use the non-poisson-picture, but shouldn't have, as this doesn't
give the strictly positive terms needed by the leastsq solvers.

This commit removes (comments out and adds a NotImplementedError)
the poissonPicture=False case of do_mlgst.  We'll leave the
plumbing in for now, in case in the future we allow for use of a
different optimization algorithm.  The line where we would
destructively truncate negative terms in the non-Poisson-picture
case is marked with a comment to this effect.