Adds "recordZeroCnts" argument to DataSet loading & construction fns.
Plumbs a connection to DataSet.add_count_dict's "recordZeroCnts"
argument to the dataset generating/filtering functions in
datasetconstruction.py and the dataset/multidataset loading functions
in loaders.py.  The default value is *always* set to True for now,
as this seems the safest option (and now it's easy to access the
alternate "sparse data" functionality).