Adds short-circuiting of sparse-data logic (optimizes code).
Adds "if firsts is not None:" conditionals around computational
logic in core.py, chi2fns.py and likelihoodfns.py so that the extra
steps needed to account for omitted probabilities are not run when
there are no such probabilities.  This also helps to identify which
code is used for dealing with omitted probabilities.