Updates default `method` in DataSet.get_degrees_of_freedom.
Changes from "all_outcomes-1" to "present_outcomes-1", as this is
the more appropriate method for many qubit data.  It's possible
this will cause problems for 1Q or 2Q data that happens to have
lots of zeros in it, but I don't think it should.  Time dependent
data is probably also not handled well here, but we'll put worrying
about that off until later.