Fixes a couple of minus-sign errors in likelihoodfns.logl
These bugs were in the "sparse data" correction part of the
logl calculation, and caused an incorrect logl value to be computed
when the data contains blanks.  This probably won't affect many
users b/c by default datasets are read in as dense; this bug was
picked up in comparions during multiQ-GST tests.