Reduces max-lengths from [1,2] to [1] in test_2Q unit test.
This test takes just under 10 minutes to run on my laptop, and is
currently timing out on TravisCI, so we'll make it a bit less intense
in hopes it finishes.  Profiling reveals there's a lot of time spent
in the onenormest() function (about 50%) in expop_multiply_prep in
matrixtools.py.  It doesn't seem so worthwhile to invsest time into
optimizing yet though, and running 2Q GST with sparse matrices and a map
calc is not typical.