Adds support for basic time-dependent GST.
"Basic" here means that that sparse data doesn't work yet, and
you can't use any penalty terms or chi2 regularization.  Getting
sparse data to work is a future priority, but compatibility with
penalty terms is probably uncessary.

Time-dependent objective functions (in objectivefns.py) for chi2
and logl have been created, and are used in do_mc2gst and do_mlgst
when they are passed `time_dependent=True` (a new arg).  These
new objective functions require fwdsim function only in the map
forward simulator currently - other fwdsims will just raise an
error. (Note: in the future we need to refactorize the additional
methods in mapforwardsim.py, some need to be Cythonized.)  Fwdsim
methods call `set_time(t)` on SPAMVecs and LinearOperators
before `.torep(...)` is called to get operator representations
at the appropriate times.  LinearOperator and SPAMVec now have "default"
set_time methods that do nothing (just "pass"), so *all* operators
and SPAM vecs can be treated as time-dependent: this allows using
a mix of time independent & dependent operators, or even all time-
independent ops, when analyzing timestamped data.

Labels within circuits now have .time members (added some time ago),
which we're now thinking should be used to hold a *relative* time
amount (i.e. a duration) for the thing being labeled.  By default
the .time of a label is 0, so "instantaneous".  In the time-dependent
forward simulation, the timestamps in a DataSet are used as
start-times, and .time values within a Circuit's labels are added
to this value as operator representations are constructed.  It may
be that we want the DataSet timestamps to be *end* times instead,
but this would require a little more work so I just did what was
easiest first.

In addition to all of this, generate_fake_data(...) has been
updated to include a `times` argument that can take a list of
timestamps, so a DataSet with "binned" time-dependent data can
easily be constructed.

Light testing has been done: performing GST with time-dependent=
True and False will give the same results when the model and data
is time-independent, as it should (except the TD case takes a lot
longer to run).